arregloMigracion

+++ database/migrations/2025_10_07_152415_create_coleccion_producto_table.php
@@ -11,17 +11,16 @@ return new class extends Migration
      */
     public function up(): void
     {
-        // Comprobamos si la tabla YA existe antes de intentar crearla
         if (!Schema::hasTable('coleccion_producto')) {
             Schema::create('coleccion_producto', function (Blueprint $table) {
-                $table->unsignedInteger('id_coleccion');
-                $table->unsignedInteger('id_producto');
+                // Definimos las columnas
+                $table->unsignedBigInteger('id_coleccion'); // BigInteger para coincidir con id()
+                $table->unsignedBigInteger('id_producto');  // BigInteger para coincidir con id()
 
-                // Definir las relaciones (Asegúrate que los nombres coinciden con tus tablas 'coleccion' y 'producto')
+                // Definimos las relaciones apuntando a 'id'
                 $table->foreign('id_coleccion')->references('id')->on('coleccion')->onDelete('cascade');
                 $table->foreign('id_producto')->references('id')->on('producto')->onDelete('cascade');
 
-                // Clave primaria compuesta
                 $table->primary(['id_coleccion', 'id_producto']);
             });
         }
